Fix reference window staying blank on first Canvas render
Even though everything seems to be the same on both renders, specifically the domElement that we are copying from, but everything else too for that matter. For now this hack will have to do, but it is too strange not to investigate further.

Also fixed an actual potential issue in ReferenceWindow where the raf loop could be permanently halted if domElement is ever missing.

Fixes #29